feat(badge): make status badge a thing
