+++ .changeset/grumpy-months-notice.md
@@ -1,0 +1,6 @@
+---
+'@twilio-paste/codemods': patch
+'@twilio-paste/core': patch
+---
+
+[Codemod]: updated exports
