+++ .changeset/twenty-mayflies-remember.md
@@ -1,0 +1,6 @@
+---
+'@twilio-paste/badge': minor
+'@twilio-paste/core': minor
+---
+
+[Badge] Add Status Badge as prescribed by the Paste Status pattern
