+++ packages/paste-codemods/tools/.cache/mappings.json
@@ -10,7 +10,14 @@
   "secureExternalLink": "@twilio-paste/core/anchor",
   "Avatar": "@twilio-paste/core/avatar",
   "Badge": "@twilio-paste/core/badge",
+  "StatusBadge": "@twilio-paste/core/badge",
+  "StatusIcons": "@twilio-paste/core/badge",
+  "StatusNames": "@twilio-paste/core/badge",
   "badgeBaseStyles": "@twilio-paste/core/badge",
+  "badgeButtonStyles": "@twilio-paste/core/badge",
+  "badgeVariantStyles": "@twilio-paste/core/badge",
+  "getBadgeAnchorStyles": "@twilio-paste/core/badge",
+  "getBadgeButtonStyles": "@twilio-paste/core/badge",
   "useResizeChildIcons": "@twilio-paste/core/badge",
   "BaseRadioCheckboxControl": "@twilio-paste/core/base-radio-checkbox",
   "BaseRadioCheckboxHelpText": "@twilio-paste/core/base-radio-checkbox",
@@ -127,10 +134,14 @@
   "OrderedList": "@twilio-paste/core/list",
   "UnorderedList": "@twilio-paste/core/list",
   "Menu": "@twilio-paste/core/menu",
+  "MenuBadge": "@twilio-paste/core/menu",
+  "MenuBadgeContent": "@twilio-paste/core/menu",
+  "MenuBadgeStatus": "@twilio-paste/core/menu",
   "MenuButton": "@twilio-paste/core/menu",
   "MenuGroup": "@twilio-paste/core/menu",
   "MenuGroupContext": "@twilio-paste/core/menu",
   "MenuItem": "@twilio-paste/core/menu",
+  "MenuItemStatus": "@twilio-paste/core/menu",
   "MenuSeparator": "@twilio-paste/core/menu",
   "StyledMenuItem": "@twilio-paste/core/menu",
   "SubMenuButton": "@twilio-paste/core/menu",
