+++ packages/paste-core/components/menu/package.json
@@ -26,6 +26,7 @@
   "peerDependencies": {
     "@twilio-paste/anchor": "^11.0.0",
     "@twilio-paste/animation-library": "^1.0.0",
+    "@twilio-paste/badge": "^7.0.0",
     "@twilio-paste/box": "^9.0.0",
     "@twilio-paste/button": "^13.0.0",
     "@twilio-paste/color-contrast-utils": "^4.0.0",
@@ -53,6 +54,7 @@
   "devDependencies": {
     "@twilio-paste/anchor": "^11.0.0",
     "@twilio-paste/animation-library": "^1.0.0",
+    "@twilio-paste/badge": "^7.0.0",
     "@twilio-paste/box": "^9.0.0",
     "@twilio-paste/button": "^13.0.0",
     "@twilio-paste/color-contrast-utils": "^4.0.0",
