+++ packages/paste-core/components/menu/src/constants.ts
@@ -1,0 +1,6 @@
+export const MenuItemVariants = {
+  DEFAULT: 'default',
+  DESTRUCTIVE: 'destructive',
+  GROUP_ITEM: 'group_item',
+  DESTRUCTIVE_GROUP_ITEM: 'destructive_group_item',
+} as const;
